+++ ft-plugin/src/main/java/com/ft/plugin/garble/FTExtension.java
@@ -27,22 +27,22 @@ public class FTExtension {
      */
     public boolean autoUploadNativeDebugSymbol = false;
     /**
-     *  datakit DCA 上传地址，一般为 datakitUrl  9531 端口
+     *  datakit 上传地址，一般为 datakitUrl  端口
      */
-    public String datakitDCAUrl = "";
+    public String datakitUrl = "";
+
     /**
-     *
+     *  dataway token
+     */
+    public String datawayToken = "";
+    /**
+     * 应用 appid
      */
     public String appId = "";
     /**
-     *
+     * 上传 sourcemap 上传环境
      */
     public String env = "prod";
-
-    /**
-     * asm 版本，默认为 asm9, 支持 asm7 ～ asm9
-     */
-    public String asmVersion = "asm9";
 
     private final NamedDomainObjectContainer<ProductFlavorModel> prodFlavor;
 
@@ -65,10 +65,10 @@ public class FTExtension {
                 ", openAutoTrack=" + openAutoTrack +
                 ", autoUploadMap=" + autoUploadMap +
                 ", autoUploadNativeDebugSymbol=" + autoUploadNativeDebugSymbol +
-                ", datakitDCAUrl='" + datakitDCAUrl + '\'' +
+                ", datakitUrl='" + datakitUrl + '\'' +
+                ", datawayToken='" + datawayToken + '\'' +
                 ", appId='" + appId + '\'' +
                 ", env='" + env + '\'' +
-                ", asmVersion='" + asmVersion + '\'' +
                 '}';
     }
 }