+++ ft-plugin/src/main/java/com/ft/plugin/garble/PluginConfigManager.java
@@ -28,7 +28,7 @@ public class PluginConfigManager {
      * @return
      */
     public int getASMVersion() {
-        String asmVersion = extension.asmVersion;
+//        String asmVersion = extension.asmVersion;
 //        if (asmVersion.equalsIgnoreCase("asm8")) {
 //            return Opcodes.ASM8;
 //        } else if (asmVersion.equalsIgnoreCase("asm9")) {